usuario/requirements.txt
Log-Api-0.0.5
SQLAlchemy-2.0.13
attrs-23.1.0 
boto3-1.26.135 
botocore-1.29.135 
greenlet-2.0.2 
jmespath-1.0.1 
jsonschema-4.17.3
pymysql-1.0.3
pyrsistent-0.19.3
python-dateutil-2.8.2
s3transfer-0.6.1
six-1.16.0
typing-extensions-4.5.0
urllib3-1.26.15